+++ mt-kahypar/datastructures/hypergraph_common.h
@@ -54,7 +54,7 @@ using HyperedgeID = uint32_t;
 // louvain graph
 using NodeID = uint32_t;
 #endif
-using HypernodeWeight = int32_t;
+using HypernodeWeight = int64_t;
 using HyperedgeWeight = int64_t;
 using PartitionID = int32_t;
 using Gain = HyperedgeWeight;
